WIP: chain rules
WIP: new idea: make several gotos between two states

Instead of keeping only the "lowest" (in the ancestors graph) label
for gotos (e.g. NUM), keep them all (expr, term, NUM) as different
gotos.

The idea is that more gotos should allow for more precise lokbacks,
hence more precise lookaheads.

We also no longer track all the FOLLOWS of a goto from the outgoing
transitions of the state: it unions too many tokens, some of them
coming from other rules.  So rather, we go back to the kernel (with
closure) of the state, and look at the tokens that can follow the
non-terminal of the goto that we consider.

WIP: partial fix (-tmp, +tests)

With this, the bison test suite passes, but not the examples of
tmp-chain: they have too many lookaheads.

WIP: yacc.c: use an assertion

* #: .