glr: also use a computed type for states
This skeleton uses a structure for states (state number, value and
location), so it is less likely to benefit from a stack size reduction
than yacc.c (which uses several stacks: state number, value and
location).

This skeleton was already using int for state numbers, so, contrary to
yacc.c, this brings nothing for large automata.

Overall, it is still nicer to make the skeletons alike.

* data/skeletons/glr.c (yyStateNum): Copy the definition from yacc.c.